Changing db.js to db.js and db.init.js
